pyproject.toml
[tool.poetry]
name = "registration"
version = "0.1.0"
description = ""
authors = ["Nik <sch-nikolaj@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
python-telegram-bot = "^13.14"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
